requirements.txt
requests
celery
requests
kafka-python